documentation added

+++ contracts/BridgeContract.sol
@@ -19,11 +19,21 @@ contract TokenBridge is Ownable {
     mapping(uint256 => bytes32) private processedSignature;
     mapping(address => ISwapToken) public supportedToken;
 
-    constructor(address _erc20, uint _chainId) {
+    /*
+     * Constructor - creates supported token and applies chainId
+     * @param {address} erc20 - Address of token
+     * @param {uint256} hainId - ChainId where contract is working
+     */
+    constructor(address _erc20, uint256 _chainId) {
         supportedToken[_erc20] = ISwapToken(_erc20);
         chainId = _chainId;
     }
-
+    
+    /**
+     * Adds or removes blockchain support by id
+     * @param {id} id - Chain id 
+     * @return {bool} - Returns true if transaction succeed
+     */
     function updateChainById(uint256 id) external onlyOwner returns(bool){
         if(supportedChainId[id]) supportedChainId[id] = false;
         else supportedChainId[id] = true;
@@ -31,6 +41,14 @@ contract TokenBridge is Ownable {
         return true;
     }
 
+    /**
+     * Sends swap transaction of token between to blockchains
+     * @param {address} to - Address of recipient
+     * @param {address} tokenAddr - Token address, that should be transfered
+     * @param {uint256} amount - Token amount
+     * @param {uint256} _chainId - ChainId of blockchain where tokens should be transfered
+     * @return {bool} - Returns true if transaction succeed
+     */
     function swap(address to, address tokenAddr, uint256 amount, uint256 _chainId) external returns(bool){
         require(address(supportedToken[tokenAddr]) != nullAddress, "Token not supported!");
         require(supportedChainId[_chainId], "Blockchain doesnt supported");
@@ -46,6 +64,18 @@ contract TokenBridge is Ownable {
         return true;
     }
 
+    /**
+     * Signs a message, if signature is valid, mints tokens to recipient
+     * @param {address} from - Address of token sender
+     * @param {address} to - Address of recipient
+     * @param {address} tokenAddr - Token address, that should be transfered
+     * @param {uint256} amount - Token amount
+     * @param {uint256} _nonce - Uniq nonce number 
+     * @param {uint8} v - v part of signature
+     * @param {bytes32} r - r part of signature
+     * @param {bytes32} s - s part of signature 
+     * @return {bool} - Returns true if transaction succeed
+     */
     function redeem(
             address from, 
             address to, 
@@ -68,16 +98,31 @@ contract TokenBridge is Ownable {
         return true;
     }
 
+    /**
+     * Concatenates prefix and message
+     * @param {bytes32} message - Message that includes swap data
+     * @return {bytes32} - Returns prefixed and hashed message
+     */
     function hashMessage(bytes32 message) private pure returns (bytes32) {
         bytes memory prefix = "\x19Ethereum Signed Message:\n32";
         return keccak256(abi.encodePacked(prefix, message));
     }
 
+    /**
+     * Adds token support for swap
+     * @param {address} tokenAddr - Address of token that should be added
+     * @return {bool} - Returns true if transaction succeed
+     */
     function includeToken(address tokenAddr) external onlyOwner returns(bool){
         supportedToken[tokenAddr] = ISwapToken(tokenAddr);
         return true;
     }
 
+    /**
+     * Deletes token support for swap
+     * @param {address} tokenAddr - Address of token that should be excluded
+     * @return {bool} - Returns true if transaction succeed
+     */
     function excludeToken(address tokenAddr) external onlyOwner returns(bool){
         supportedToken[tokenAddr] = ISwapToken(nullAddress);
         return true;
